+++ gateway/src/main/proto/com/improving/app/gateway/domain/member/memberGatewayDomainEvents.proto
@@ -1,0 +1,43 @@
+syntax =  "proto3";
+
+package com.improving.app.gateway.domain.member;
+
+import "com/improving/app/common/domain/address.proto";
+import "com/improving/app/common/domain/ids.proto";
+import "com/improving/app/gateway/domain/member/memberGatewayDomainObjects.proto";
+import "google/protobuf/timestamp.proto";
+import "scalapb/scalapb.proto";
+import "scalapb/validate.proto";
+import "validate/validate.proto";
+
+message MemberRegistered {
+  common.domain.MemberId member_id = 1;
+  EditableMemberInfo member_info = 2;
+  MemberMetaInfo meta = 3;
+}
+
+message MemberEvent{
+
+  option (scalapb.message).sealed_oneof_extends = "com.improving.app.common.serialize.PBMsgOneOfSerializable";
+
+  oneof sealed_value {
+    MemberRegistered member_registered_value = 1;
+  }
+}
+
+message MemberData {
+  common.domain.MemberId member_id = 1;
+  MemberInfo member_info = 2;
+  MemberMetaInfo member_meta_info = 3;
+}
+
+message MemberEventResponse {
+  MemberEvent member_event = 1;
+}
+
+message MemberResponse {
+  oneof sealed_value {
+    MemberData member_state_value = 1;
+    MemberEventResponse member_event_value = 2;
+  }
+}
