+++ gateway/src/main/scala/com/improving/app/gateway/domain/common/util.scala
@@ -1,20 +1,18 @@
 package com.improving.app.gateway.domain.common
 
-import com.improving.app.member.domain.{
-  EditableInfo => EditableMemberInfo,
-  MemberInfo,
-  MemberMetaInfo,
-  MemberState,
-  NotificationPreference
-}
-import com.typesafe.config.ConfigFactory
-import com.improving.app.gateway.domain.{
-  MemberInfo => GatewayMemberInfo,
+import com.improving.app.gateway.domain.member.{
   EditableMemberInfo => GatewayEditableMemberInfo,
   MemberMetaInfo => GatewayMemberMetaInfo,
   MemberStates => GatewayMemberStates,
   NotificationPreference => GatewayNotificationPreference
 }
+import com.improving.app.member.domain.{
+  MemberMetaInfo,
+  MemberState,
+  NotificationPreference,
+  EditableInfo => EditableMemberInfo
+}
+import com.typesafe.config.ConfigFactory
 
 object util {
 
@@ -26,62 +24,5 @@ object util {
       config.getString(s"services.$serviceName.host"),
       config.getInt(s"services.$serviceName.port")
     )
-  }
-
-  def gatewayEditableMemberInfoToEditableInfo(info: GatewayEditableMemberInfo): EditableMemberInfo = EditableMemberInfo(
-    info.handle,
-    info.avatarUrl,
-    info.firstName,
-    info.lastName,
-    info.notificationPreference.map(gatewayNotificationPreferenceToNotificationPreference),
-    info.contact,
-    info.organizationMembership,
-    info.tenant
-  )
-
-  def editableMemberInfoToGatewayEditableInfo(info: EditableMemberInfo): GatewayEditableMemberInfo =
-    GatewayEditableMemberInfo(
-      info.handle,
-      info.avatarUrl,
-      info.firstName,
-      info.lastName,
-      info.notificationPreference.map(notificationPreferenceToGatewayNotificationPreference),
-      info.contact,
-      info.organizationMembership,
-      info.tenant
-    )
-
-  def memberStateToGatewayMemberState(memberStatus: MemberState): GatewayMemberStates = {
-    if (memberStatus.isMemberStateDraft) GatewayMemberStates.MEMBER_STATES_DRAFT
-    else if (memberStatus.isMemberStateActive) GatewayMemberStates.MEMBER_STATES_ACTIVE
-    else GatewayMemberStates.MEMBER_STATES_SUSPENDED
-  }
-
-  def memberMetaToGatewayMemberMeta(meta: MemberMetaInfo): GatewayMemberMetaInfo = GatewayMemberMetaInfo(
-    meta.createdOn,
-    meta.createdBy,
-    meta.lastModifiedOn,
-    meta.lastModifiedBy,
-    memberStateToGatewayMemberState(meta.currentState)
-  )
-
-  def notificationPreferenceToGatewayNotificationPreference(
-      notificationPreference: NotificationPreference
-  ): GatewayNotificationPreference =
-    if (notificationPreference.isNotificationPreferenceEmail)
-      GatewayNotificationPreference.NOTIFICATION_PREFERENCE_EMAIL
-    else if (notificationPreference.isNotificationPreferenceSms)
-      GatewayNotificationPreference.NOTIFICATION_PREFERENCE_SMS
-    else GatewayNotificationPreference.NOTIFICATION_PREFERENCE_APPLICATION
-
-  def gatewayNotificationPreferenceToNotificationPreference(
-      notificationPreference: GatewayNotificationPreference
-  ): NotificationPreference = notificationPreference match {
-    case GatewayNotificationPreference.NOTIFICATION_PREFERENCE_EMAIL =>
-      NotificationPreference.NOTIFICATION_PREFERENCE_EMAIL
-    case GatewayNotificationPreference.NOTIFICATION_PREFERENCE_SMS =>
-      NotificationPreference.NOTIFICATION_PREFERENCE_SMS
-    case _ =>
-      NotificationPreference.NOTIFICATION_PREFERENCE_APPLICATION
   }
 }
